refactor: useBookmark 훅으로 북마크 업데이트 로직 분리
